Fixed 3rd party css import

+++ app/javascript/bundles/RestaurantsApp/components/RestaurantCard.jsx
@@ -28,7 +28,7 @@ export default class RestaurantCard extends React.Component {
                         <div>{restaurant.address}</div>
                         {restaurant.speed && (<div>{restaurant.speed} minute delivery</div>)}
                         {restaurant.kosher === true && (<div>Kosher</div>)}
-                        {restaurant.rating && (<StarsRating rating={parseInt(restaurant.rating)}/>)}
+                        {restaurant.rating && (<StarsRating rating={parseFloat(restaurant.rating)}/>)}
                         {restaurant.accepts_10bis && (
                             <img src="assets/10bis.png" className="icon-10-bis restaurant-icon"/>
                         )}
